Refactor bench 2669628
