+++ src/search.rs
@@ -33,7 +33,7 @@ pub fn search(board: &Board, td: &mut ThreadData) -> (Move, i32) {
     td.pv.clear(0);
     td.nnue.activate(board);
 
-    let root_moves = movegen::gen_legal_moves(board);
+    let root_moves = board.gen_legal_moves();
     match root_moves.len {
         0 => return handle_no_legal_moves(board, td),
         1 => return handle_one_legal_move(board, td, &root_moves),
