+++ src/tools/scharnagl.rs
@@ -1,11 +1,11 @@
 use crate::board::castling::Rights;
+use crate::board::file::File;
 use crate::board::piece::Piece;
+use crate::board::rank::Rank;
 use crate::board::side::Side;
 use crate::board::square::Square;
 use crate::board::zobrist::Zobrist;
 use crate::board::Board;
-use crate::board::file::File;
-use crate::board::rank::Rank;
 
 /// Derive the starting positions of the backrank pieces from a scharnagl index. Used for creating
 /// a starting position for FRC and DFRC chess. Implementation based on Viridithas. Original source:
