+++ src/tools/uci.rs
@@ -1,4 +1,4 @@
-use crate::board::movegen::{gen_moves, MoveFilter};
+use crate::board::movegen::MoveFilter;
 use crate::board::moves::Move;
 use crate::board::side::Side::{Black, White};
 use crate::board::Board;
@@ -220,7 +220,7 @@ impl UCI {
         self.td.keys.push(self.board.hash());
 
         moves.iter().for_each(|m| {
-            let mut legal_moves = gen_moves(&self.board, MoveFilter::All);
+            let mut legal_moves = self.board.gen_moves(MoveFilter::All);
             let legal_move = legal_moves.iter()
                 .map(|entry| entry.mv)
                 .find(|lm| lm.matches(m));
